+++ src/interfaces/chat/ChatInfo.interface.ts
@@ -1,5 +1,5 @@
-import type { Chat } from './Chat.interface';
-import type { User } from './User.interface';
+import type { Chat } from '@/interfaces/chat/Chat.interface';
+import type { User } from '@/interfaces/user/User.interface';
 
 export interface ChatInfo {
   id: number;
