+++ src/interfaces/game/GameHistory.interface.ts
