+++ src/interfaces/user/UserGameHistory.interface.ts
